+++ json-web-token/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.0.4
+
+## Changes
+* `JwtClaimsSetKey` performance improved by caching lower-case claim-names and never throwing `Exception`s for flow
+    control : [COMMONS-76](https://bugster.forgerock.org/jira/browse/COMMONS-76)
+
 # 3.0.3
 
 ## Changes